+++ scheduler/tests/test_internals.py
@@ -11,7 +11,7 @@ from scheduler.tests.testtools import SchedulerBaseCase, task_factory
 
 class TestInternals(SchedulerBaseCase):
     def test_get_scheduled_job(self):
-        task = task_factory(TaskType.ONCE, scheduled_time=timezone.now() - timedelta(hours=1))
+        task = task_factory(TaskType.ONCE, scheduled_time=timezone.now() + timedelta(hours=1))
         self.assertEqual(task, get_scheduled_task(TaskType.ONCE, task.id))
         with self.assertRaises(ValueError):
             get_scheduled_task(task.task_type, task.id + 1)