+++ scheduler/tests/test_task_types/test_repeatable_task.py
@@ -35,6 +35,7 @@ class TestRepeatableTask(BaseTestCases.TestSchedulableTask):
         self.assertEqual(task.interval_unit, "seconds")
 
     def test_create_task_without_scheduled_time(self):
+        settings.SCHEDULER_CONFIG.SCHEDULER_INTERVAL = 10
         task = Task.objects.create(
             name="konichiva_every_2s",
             callable="scheduler.tests.jobs.test_args_kwargs",
@@ -52,44 +53,44 @@ class TestRepeatableTask(BaseTestCases.TestSchedulableTask):
         self.assertEqual(task.interval_unit, "seconds")
 
     def test_unschedulable_old_job(self):
-        job = task_factory(self.task_type, scheduled_time=timezone.now() - timedelta(hours=1), repeat=0)
-        self.assertFalse(job.is_scheduled())
+        task = task_factory(TaskType.REPEATABLE, scheduled_time=timezone.now() - timedelta(hours=1), repeat=0)
+        self.assertFalse(task.is_scheduled())
 
     def test_schedulable_old_job_repeat_none(self):
         # If repeat is None, the job should be scheduled
-        job = task_factory(self.task_type, scheduled_time=timezone.now() - timedelta(hours=1), repeat=None)
-        self.assertTrue(job.is_scheduled())
+        task = task_factory(TaskType.REPEATABLE, scheduled_time=timezone.now() - timedelta(hours=1), repeat=None)
+        self.assertTrue(task.is_scheduled())
 
     def test_clean(self):
-        job = task_factory(self.task_type)
-        job.queue = self.queue_name
-        job.callable = "scheduler.tests.jobs.test_job"
-        job.interval = 1
-        job.success_ttl = -1
-        self.assertIsNone(job.clean())
+        task = task_factory(TaskType.REPEATABLE)
+        task.queue = self.queue_name
+        task.callable = "scheduler.tests.jobs.test_job"
+        task.interval = 1
+        task.success_ttl = -1
+        self.assertIsNone(task.clean())
 
     def test_clean_seconds(self):
-        job = task_factory(self.task_type)
-        job.queue = self.queue_name
-        job.callable = "scheduler.tests.jobs.test_job"
-        job.interval = 60
-        job.success_ttl = -1
-        job.interval_unit = "seconds"
-        self.assertIsNone(job.clean())
+        task = task_factory(TaskType.REPEATABLE)
+        task.queue = self.queue_name
+        task.callable = "scheduler.tests.jobs.test_job"
+        task.interval = 60
+        task.success_ttl = -1
+        task.interval_unit = "seconds"
+        task.clean()
 
     @override_settings(SCHEDULER_CONFIG=SchedulerConfiguration(SCHEDULER_INTERVAL=10))
     def test_clean_too_frequent(self):
-        job = task_factory(self.task_type)
-        job.queue = self.queue_name
-        job.callable = "scheduler.tests.jobs.test_job"
-        job.interval = 2  # Smaller than 10
-        job.success_ttl = -1
-        job.interval_unit = "seconds"
+        task = task_factory(TaskType.REPEATABLE)
+        task.queue = self.queue_name
+        task.callable = "scheduler.tests.jobs.test_job"
+        task.interval = 2  # Smaller than 10
+        task.success_ttl = -1
+        task.interval_unit = "seconds"
         with self.assertRaises(ValidationError):
-            job.clean_interval_unit()
+            task.clean_interval_unit()
 
     def test_clean_short_result_ttl(self):
-        task = task_factory(self.task_type)
+        task = task_factory(TaskType.REPEATABLE)
         task.queue = self.queue_name
         task.callable = "scheduler.tests.jobs.test_job"
         task.interval = 1
@@ -101,7 +102,7 @@ class TestRepeatableTask(BaseTestCases.TestSchedulableTask):
             task.clean_result_ttl()
 
     def test_clean_indefinite_result_ttl(self):
-        task = task_factory(self.task_type)
+        task = task_factory(TaskType.REPEATABLE)
         task.queue = self.queue_name
         task.callable = "scheduler.tests.jobs.test_job"
         task.interval = 1
@@ -110,7 +111,7 @@ class TestRepeatableTask(BaseTestCases.TestSchedulableTask):
         task.clean_result_ttl()
 
     def test_clean_undefined_result_ttl(self):
-        task = task_factory(self.task_type)
+        task = task_factory(TaskType.REPEATABLE)
         task.queue = self.queue_name
         task.callable = "scheduler.tests.jobs.test_job"
         task.interval = 1
@@ -118,56 +119,56 @@ class TestRepeatableTask(BaseTestCases.TestSchedulableTask):
         task.clean_result_ttl()
 
     def test_interval_seconds_weeks(self):
-        task = task_factory(self.task_type, interval=2, interval_unit="weeks")
+        task = task_factory(TaskType.REPEATABLE, interval=2, interval_unit="weeks")
         self.assertEqual(1209600.0, task.interval_seconds())
 
     def test_interval_seconds_days(self):
-        task = task_factory(self.task_type, interval=2, interval_unit="days")
+        task = task_factory(TaskType.REPEATABLE, interval=2, interval_unit="days")
         self.assertEqual(172800.0, task.interval_seconds())
 
     def test_interval_seconds_hours(self):
-        job = task_factory(self.task_type, interval=2, interval_unit="hours")
-        self.assertEqual(7200.0, job.interval_seconds())
+        task = task_factory(TaskType.REPEATABLE, interval=2, interval_unit="hours")
+        self.assertEqual(7200.0, task.interval_seconds())
 
     def test_interval_seconds_minutes(self):
-        job = task_factory(self.task_type, interval=15, interval_unit="minutes")
-        self.assertEqual(900.0, job.interval_seconds())
+        task = task_factory(TaskType.REPEATABLE, interval=15, interval_unit="minutes")
+        self.assertEqual(900.0, task.interval_seconds())
 
     def test_interval_seconds_seconds(self):
-        job = task_factory(self.task_type, interval=15, interval_unit="seconds")
-        self.assertEqual(15.0, job.interval_seconds())
+        task = task_factory(TaskType.REPEATABLE, interval=15, interval_unit="seconds")
+        self.assertEqual(15.0, task.interval_seconds())
 
     def test_result_interval(self):
-        job = task_factory(self.task_type)
-        entry = _get_task_scheduled_job_from_registry(job)
+        task = task_factory(TaskType.REPEATABLE)
+        entry = _get_task_scheduled_job_from_registry(task)
         self.assertEqual(entry.meta["interval"], 3600)
 
     def test_repeat(self):
-        job = task_factory(self.task_type, repeat=10)
-        entry = _get_task_scheduled_job_from_registry(job)
+        task = task_factory(TaskType.REPEATABLE, repeat=10)
+        entry = _get_task_scheduled_job_from_registry(task)
         self.assertEqual(entry.meta["repeat"], 10)
 
     def test_repeat_old_job_exhausted(self):
         base_time = timezone.now()
-        job = task_factory(self.task_type, scheduled_time=base_time - timedelta(hours=10), repeat=10)
-        self.assertEqual(job.is_scheduled(), False)
+        task = task_factory(TaskType.REPEATABLE, scheduled_time=base_time - timedelta(hours=10), repeat=10)
+        self.assertEqual(task.is_scheduled(), False)
 
     def test_repeat_old_job_last_iter(self):
         base_time = timezone.now()
-        job = task_factory(self.task_type, scheduled_time=base_time - timedelta(hours=9, minutes=30), repeat=10)
-        self.assertEqual(job.repeat, 0)
-        self.assertEqual(job.is_scheduled(), True)
+        task = task_factory(TaskType.REPEATABLE, scheduled_time=base_time - timedelta(hours=9, minutes=30), repeat=10)
+        self.assertEqual(task.repeat, 0)
+        self.assertEqual(task.is_scheduled(), True)
 
     def test_repeat_old_job_remaining(self):
         base_time = timezone.now()
-        job = task_factory(self.task_type, scheduled_time=base_time - timedelta(minutes=30), repeat=5)
-        self.assertEqual(job.repeat, 4)
-        self.assertEqual(job.scheduled_time, base_time + timedelta(minutes=30))
-        self.assertEqual(job.is_scheduled(), True)
+        task = task_factory(TaskType.REPEATABLE, scheduled_time=base_time - timedelta(minutes=30), repeat=5)
+        self.assertEqual(task.repeat, 4)
+        self.assertEqual(task.scheduled_time, base_time + timedelta(minutes=30))
+        self.assertEqual(task.is_scheduled(), True)
 
     def test_repeat_none_interval_2_min(self):
         base_time = timezone.now()
-        task = task_factory(self.task_type, scheduled_time=base_time - timedelta(minutes=29), repeat=None)
+        task = task_factory(TaskType.REPEATABLE, scheduled_time=base_time - timedelta(minutes=29), repeat=None)
         task.interval = 120
         task.interval_unit = "seconds"
         task.save()
@@ -175,7 +176,7 @@ class TestRepeatableTask(BaseTestCases.TestSchedulableTask):
         self.assertTrue(task.is_scheduled())
 
     def test_check_rescheduled_after_execution(self):
-        task = task_factory(self.task_type, scheduled_time=timezone.now() + timedelta(seconds=1), repeat=10)
+        task = task_factory(TaskType.REPEATABLE, scheduled_time=timezone.now() + timedelta(seconds=1), repeat=10)
         queue = task.rqueue
         first_run_id = task.job_name
         entry = JobModel.get(first_run_id, connection=queue.connection)
@@ -190,7 +191,7 @@ class TestRepeatableTask(BaseTestCases.TestSchedulableTask):
 
     def test_check_rescheduled_after_execution_failed_job(self):
         task = task_factory(
-            self.task_type,
+            TaskType.REPEATABLE,
             callable_name="scheduler.tests.jobs.failing_job",
             scheduled_time=timezone.now() + timedelta(seconds=1),
             repeat=10,
@@ -209,7 +210,7 @@ class TestRepeatableTask(BaseTestCases.TestSchedulableTask):
 
     def test_check_not_rescheduled_after_last_repeat(self):
         task = task_factory(
-            self.task_type,
+            TaskType.REPEATABLE,
             scheduled_time=timezone.now() + timedelta(seconds=1),
             repeat=1,
         )